graybox refactor
