+++ tools/floodbox/graybox/graybox.css
@@ -1,14 +1,3 @@
-:host {
-  display: block;
-  max-width: var(--grid-container-width);
-  margin: var(--spacing-500);
-}
-
-body {
-  background-color: #fff;
-  margin: 0 var(--spacing-1000);
-}
-
 .info-box {
   background: var(--s2-gray-75);
   border-radius: var(--s2-radius-100);
