+++ tools/graybox.html
@@ -5,6 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1"/>
     <link rel="icon" href="data:,">
     <link rel="stylesheet" href="https://da.live/nx/styles/nexter.css">
+    <link rel="stylesheet" href="/tools/floodbox/floodbox.css">
     <link rel="stylesheet" href="/tools/floodbox/graybox/graybox.css">
     <!-- Import DA App SDK -->
     <script src="https://da.live/nx/utils/sdk.js" type="module"></script>
